requirements.txt
asgiref==3.8.1
attrs==25.1.0
autobahn==24.4.2
Automat==24.8.1
cffi==1.17.1
channels==4.2.0
constantly==23.10.4
cryptography==44.0.0
daphne==4.1.2
diff-match-patch==20241021
Django==5.1.5
django-ckeditor-5==0.2.15
django-cors-headers==4.6.0
django-filter==24.3
django-import-export==4.3.4
django-jazzmin==3.0.1
django-ranged-response==0.2.0
django-simple-captcha==0.6.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-yasg==1.21.8
gunicorn==23.0.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
packaging==24.2
pillow==11.1.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.10.1
pyOpenSSL==25.0.0
pytz==2024.2
PyYAML==6.0.2
service-identity==24.2.0
sqlparse==0.5.3
tablib==3.8.0
tomli==2.2.1
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
zope.interface==7.2
